+++ core/core-frontend/src/views/template-market/component/CategoryTemplateV2.vue
